Move pycodestyle violations to rule modules
This doesn't change where `IOError`s come from though, that felt unintuitive.
