+++ src/rules/pycodestyle/rules/mod.rs
@@ -4,6 +4,11 @@ pub use ambiguous_variable_name::{ambiguous_variable_name, AmbiguousVariableName
 pub use do_not_assign_lambda::{do_not_assign_lambda, DoNotAssignLambda};
 pub use do_not_use_bare_except::{do_not_use_bare_except, DoNotUseBareExcept};
 pub use doc_line_too_long::{doc_line_too_long, DocLineTooLong};
+pub use errors::{syntax_error, IOError, SyntaxError};
+pub use imports::{
+    module_import_not_at_top_of_file, multiple_imports_on_one_line, ModuleImportNotAtTopOfFile,
+    MultipleImportsOnOneLine,
+};
 pub use invalid_escape_sequence::{invalid_escape_sequence, InvalidEscapeSequence};
 pub use line_too_long::{line_too_long, LineTooLong};
 pub use literal_comparisons::{literal_comparisons, NoneComparison, TrueFalseComparison};
@@ -18,6 +23,8 @@ mod ambiguous_variable_name;
 mod do_not_assign_lambda;
 mod do_not_use_bare_except;
 mod doc_line_too_long;
+mod errors;
+mod imports;
 mod invalid_escape_sequence;
 mod line_too_long;
 mod literal_comparisons;
